Modifed the data model-SNLProject-2094

+++ Smps.Infrastructure/User.cs
@@ -24,7 +24,7 @@ namespace Smps.DAL
         public int EmpNo { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
-        public Nullable<int> MobileNumber { get; set; }
+        public Nullable<long> MobileNumber { get; set; }
         public string UserLoginId { get; set; }
         public string UserLoginPassword { get; set; }
         public string UserType { get; set; }
